benchmark/requirements.txt
accelerate
bitsandbytes
datasets
lxml
matplotlib
numpy
omegaconf
pandas
peft
pyyaml
pytest
seaborn
torch
transformers
trl
